Related work items: #19047, #19282, #19300, #19306, #19405

+++ dotnet/corecomponents/net.atos.daf.ct2.group/repository/GroupRepository.cs
@@ -101,34 +101,69 @@ namespace net.atos.daf.ct2.group
             return group;
         }
 
-        public async Task<bool> Delete(long groupid, ObjectType objectType)
+        public async Task<bool> CanDelete(long groupId, ObjectType objectType)
         {
             try
             {
                 var parameter = new DynamicParameters();
-                parameter.Add("@id", groupid);
+                parameter.Add("@id", groupId);
                 string query = string.Empty;
-                //TODO: Need to prepare this as single for delete all ref. of group
-                using (var transactionScope = new TransactionScope(TransactionScopeAsyncFlowOption.Enabled))
+
+                if (objectType == ObjectType.AccountGroup)
+                    query = @"select exists(select 1 from master.accessrelationship where account_group_id = @id)";
+                else
                 {
-                    // delete access relation ship
-                    if (objectType == ObjectType.AccountGroup)
-                        query = @"delete from master.accessrelationship where account_group_id = @id";
-                    else query = @"delete from master.accessrelationship where vehicle_group_id = @id";
-                    await _dataAccess.ExecuteScalarAsync<int>(query, parameter);
-
-
-                    // delete group ref
-                    query = @"delete from master.groupref where group_id = @id";
-                    await _dataAccess.ExecuteScalarAsync<int>(query, parameter);
-
-                    // delete group 
-                    query = @"delete from master.group where id = @id";
-                    await _dataAccess.ExecuteScalarAsync<int>(query, parameter);
-                    transactionScope.Complete();
-                    return true;
+                    // Check for existing Access relationship, Alert and Org relationship
+                    // TODO - Scheduler check
+                    query = @"select exists
+                              (
+	                              select 1 from master.accessrelationship where vehicle_group_id = @id
+	                              UNION
+	                              select 1 from master.alert where vehicle_group_id = @id and state in ('A', 'I')
+                                  UNION
+	                              select 1
+	                              from master.orgrelationshipmapping orm
+                                  inner join master.orgrelationship ors on orm.relationship_id=ors.id and orm.vehicle_group_id = @id and ors.state='A' 
+	                              where case when COALESCE(end_date,0) !=0 then to_timestamp(COALESCE(end_date)/1000)::date>now()::date
+	                                    else COALESCE(end_date,0) = 0 end
+                              )";
                 }
 
+                return !await _dataAccess.ExecuteScalarAsync<bool>(query, parameter);
+            }
+            catch (Exception)
+            {
+                throw;
+            }
+        }
+
+        public async Task<VehicleGroupDelete> Delete(long groupId, ObjectType objectType)
+        {
+            var response = new VehicleGroupDelete();
+            try
+            {
+                response.CanDelete = await CanDelete(groupId, objectType);
+                if (response.CanDelete)
+                {
+                    var parameter = new DynamicParameters();
+                    parameter.Add("@id", groupId);
+                    string query = string.Empty;
+                    //TODO: Need to prepare this as single for delete all ref. of group
+                    using (var transactionScope = new TransactionScope(TransactionScopeAsyncFlowOption.Enabled))
+                    {
+                        // delete group ref
+                        query = @"delete from master.groupref where group_id = @id";
+                        await _dataAccess.ExecuteScalarAsync<int>(query, parameter);
+
+                        // delete group 
+                        query = @"delete from master.group where id = @id";
+                        await _dataAccess.ExecuteScalarAsync<int>(query, parameter);
+                        transactionScope.Complete();
+
+                        response.IsDeleted = true;
+                    }
+                }
+                return response;
             }
             catch (Exception)
             {
